llm-room/requirements.txt
fastapi
uvicorn[standard]
websockets
langchain
langchain-openai
pyyaml
python-dotenv
motor
httpx
